Rage::join in use instead of the old join.

+++ src/GameCommand.cpp
@@ -476,8 +476,8 @@ void GameCommand::LoadOne( const Command& cmd )
 		if( cmd.m_vsArgs.size() == 3 )
 		{
 			m_bFadeMusic = true;
-			m_fMusicFadeOutVolume = static_cast<float>(atof( cmd.m_vsArgs[1] ));
-			m_fMusicFadeOutSeconds = static_cast<float>(atof( cmd.m_vsArgs[2] ));
+      m_fMusicFadeOutVolume = static_cast<float>(std::stof( cmd.m_vsArgs[1] ));
+      m_fMusicFadeOutSeconds = static_cast<float>(std::stof( cmd.m_vsArgs[2] ));
 		}
 		else
 		{
